+++ db/src/main/java/com/sonicbase/query/impl/GroupByContext.java
@@ -14,6 +14,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class GroupByContext {
   private List<FieldContext> fieldContexts;
@@ -153,7 +154,9 @@ public class GroupByContext {
           int len = (int) DataUtil.readVLong(in);
           byte[] buffer = new byte[len];
           in.readFully(buffer);
-          groupValues[k] = DatabaseCommon.deserializeFields(dbName, common, buffer, 0, fieldContexts.get(k).tableSchema, common.getSchemaVersion(), null, new AtomicInteger(), true)[0];
+          groupValues[k] = DatabaseCommon.deserializeFields(dbName, common, buffer, 0,
+              fieldContexts.get(k).tableSchema, common.getSchemaVersion(), null, new AtomicLong(),
+              true)[0];
         }
         Counter counter = new Counter();
         counter.deserialize(in);
@@ -202,7 +205,8 @@ public class GroupByContext {
       out.writeInt(innerMap.getValue().values().size());
       for (GroupCounter counter : innerMap.getValue().values()) {
         for  (int i = 0; i < count; i++) {
-          DatabaseCommon.serializeFields(new Object[]{counter.groupValues[i]}, out, fieldContexts.get(i).tableSchema, common.getSchemaVersion(), true);
+          DatabaseCommon.serializeFields(new Object[]{counter.groupValues[i]}, out, fieldContexts.get(i).tableSchema,
+              common.getSchemaVersion(), true);
         }
         out.write(counter.counter.serialize());
       }
